+++ frontend/src/app/services/users.service.ts
@@ -42,13 +42,21 @@ export class UsersService {
       if(ck!=null){
         var tk = JSON.parse(ck);
         var res = [];
+        let isMedic = false;
         for(var i in tk){
+          if(i==='medic id'){
+            console.log('medico');
+            isMedic = true;
+          }
           res.push(tk[i]);
         }
         let headers = new HttpHeaders()
         headers = headers.set('Authorization', 'Token '+res[0])
-
-        return this.http.get(API_URL + `users/patients/${res[1]}/`,{'headers':headers})
+        if(isMedic){
+          return this.http.get(API_URL + `users/medics/${res[1]}/`,{'headers':headers})
+        }else{
+          return this.http.get(API_URL + `users/patients/${res[1]}/`,{'headers':headers})
+        }
       }
     }
     return new Observable<any>;
