+++ frontend/src/app/tabs/tabs.page.html
@@ -28,13 +28,6 @@
       </ion-tab-button>
     </ion-tab-bar>
   </div>
-  <div *ngIf="!isPremium" slot="bottom">
-    <br>
-    <br>
-    <br>
-    <br>
-    <br>
-  </div>
 
 </ion-tabs>
